download windows pre-req files using the host OS
Many pre-req file have moved to require TLS 1.2, to address this
downloading these files in the host system will allows this requirement
to be meet even when the guest OS being built does not yet support TLS 1.2.

* downloads dotnet and wmf from a provisioner run by the host OS
* updates dotnet to 4.5.2
* update wmf for powershell 5.1
* adjust install process for wmf to run as SYSTEM
* moves boxstarter and Ruby install into the chocolatey_installs path

Furhter work should be done to validate the downlaod hash and support caching downloads
instead of force each build to download again.

This does not remove the need for downloads by chocolatey when preforming those installs.